new errors
Looks like there's an element in the index.js file, specifically the app element on line 11, which is "not a DOM element" according to the browser element examiner. This looks like a package dependency issue to me, this book must be getting old because it looks like DOM is unsupported.